requirements.txt
flask==3.0.0
werkzeug==3.0.1
openai>=1.0.0
python-dotenv==0.19.0
googlemaps==4.10.0